Additional Mono rendering fixes
- on initial rendering mono likes to send a resize event where the width
is 0, ignore it.  This was causing gbDatInfo to not render until a
manual resize was done.

- set AutoScaleMode to System.Windows.Forms.AutoScaleMode.None.  This
fixes a number of rendering issues under mono and didnt seem to have any
negative effect under c#/windows builds.

- tweak some of the initial form sizes and locations on FrmMain.  A
number of the setting didnt make sense based on the size of the panel
they were put it.  This resolves an issue with the gbDatInfo and
gbSetInfo starting off with the wrong size under mono.